Fix NPE with unresponsive Weaviate server (#86)
* Add unit test to reproduce NPE

An unresponsive weaviate server will cause a NPE and the unit test
added here reproduces that NPE.

* fix #85 throw WeaviateResultError instead of NPE

The weaviate client will ocasionally return null when calling
result.getResult. This can happen when the Weaviate server is not
accessible or when the Weaviate server is overloaded and not able to
return a response temporarily.